+++ apps/frontend/app/admin/contest/[id]/(overall)/@tabs/page.tsx
@@ -1,0 +1,3 @@
+export default function Participant() {
+  return <>Participant</>
+}
