+++ apps/frontend/app/admin/contest/[id]/(overall)/@tabs/submission/page.tsx
@@ -1,0 +1,3 @@
+export default function Submission() {
+  return <>Submission</>
+}
