+++ apps/frontend/app/admin/contest/[id]/(overall)/error.tsx
@@ -1,0 +1,20 @@
+'use client'
+
+import ErrorDetail from '@/components/ErrorDetail'
+import { captureException } from '@sentry/nextjs'
+import { useEffect } from 'react'
+
+interface Props {
+  error: Error & { digest?: string }
+  reset: () => void
+}
+
+export default function Error({ error }: Props) {
+  useEffect(() => {
+    captureException(error)
+  }, [error])
+
+  return (
+    <ErrorDetail errorDetail="Failed to Load Contest Overall!" error={error} />
+  )
+}
