+++ apps/frontend/app/admin/contest/[id]/edit/error.tsx
