+++ apps/frontend/app/admin/contest/create/page.tsx
@@ -1,17 +1,8 @@
 'use client'
 
 import { DataTableAdmin } from '@/components/DataTableAdmin'
-import TextEditor from '@/components/TextEditor'
-import { DateTimePickerDemo } from '@/components/date-time-picker-demo'
 import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger
-} from '@/components/ui/popover'
 import { ScrollArea } from '@/components/ui/scroll-area'
-import { Switch } from '@/components/ui/switch'
 import {
   CREATE_CONTEST,
   IMPORT_PROBLEMS_TO_CONTEST
@@ -20,43 +11,25 @@ import {
   UPDATE_PROBLEM_VISIBLE,
   UPDATE_CONTEST_PROBLEMS_ORDER
 } from '@/graphql/problem/mutations'
-import { cn } from '@/lib/utils'
 import { useMutation } from '@apollo/client'
 import type { CreateContestInput } from '@generated/graphql'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { PlusCircleIcon } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
-import { useForm, Controller } from 'react-hook-form'
+import { useForm, FormProvider } from 'react-hook-form'
 import { FaAngleLeft } from 'react-icons/fa6'
 import { IoMdCheckmarkCircleOutline } from 'react-icons/io'
-import { MdHelpOutline } from 'react-icons/md'
-import { PiWarningBold } from 'react-icons/pi'
 import { toast } from 'sonner'
-import { z } from 'zod'
-import Label from '../_components/Label'
+import DescriptionForm from '../../_components/DescriptionForm'
+import FormSection from '../../_components/FormSection'
+import SwitchField from '../../_components/SwitchField'
+import TitleForm from '../../_components/TitleForm'
+import ContestProblemListLabel from '../_components/ContestProblemListLabel'
+import ImportProblemButton from '../_components/ImportProblemButton'
+import TimeForm from '../_components/TimeForm'
+import { type ContestProblem, createSchema } from '../utils'
 import { columns } from './_components/Columns'
-
-const inputStyle =
-  'border-gray-200 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-gray-950'
-
-const schema = z.object({
-  title: z.string().min(1).max(100),
-  isRankVisible: z.boolean().optional(),
-  isVisible: z.boolean().optional(),
-  description: z.string().min(1),
-  startTime: z.date(),
-  endTime: z.date(),
-  invitationCode: z.string().min(6).max(6).nullish()
-})
-
-interface ContestProblem {
-  id: number
-  title: string
-  order: number
-  difficulty: string
-}
 
 export default function Page() {
   const [problems, setProblems] = useState<ContestProblem[]>([])
@@ -66,20 +39,15 @@ export default function Page() {
 
   const router = useRouter()
 
-  const {
-    handleSubmit,
-    control,
-    register,
-    getValues,
-    setValue,
-    formState: { errors }
-  } = useForm<CreateContestInput>({
-    resolver: zodResolver(schema),
+  const methods = useForm<CreateContestInput>({
+    resolver: zodResolver(createSchema),
     defaultValues: {
       isRankVisible: true,
       isVisible: true
     }
   })
+
+  const { handleSubmit, getValues, setValue } = methods
 
   const [createContest, { error }] = useMutation(CREATE_CONTEST)
   const [importProblemsToContest] = useMutation(IMPORT_PROBLEMS_TO_CONTEST)
@@ -213,195 +181,56 @@ export default function Page() {
           onSubmit={handleSubmit(onSubmit)}
           className="flex w-[760px] flex-col gap-6"
         >
-          <div className="flex gap-6">
-            <div className="flex flex-col gap-1">
-              <Label>Title</Label>
-              <Input
-                id="title"
-                type="text"
-                placeholder="Name your contest"
-                className={cn(inputStyle, 'w-[380px]')}
-                {...register('title')}
-              />
-              {errors.title && (
-                <div className="flex items-center gap-1 text-xs text-red-500">
-                  <PiWarningBold />
-                  {getValues('title').length === 0
-                    ? 'Required'
-                    : errors.title.message}
-                </div>
+          <FormProvider {...methods}>
+            <FormSection title="Title">
+              <TitleForm placeholder="Name your contest" />
+            </FormSection>
+            <div className="flex gap-6">
+              <FormSection title="Start Time">
+                <TimeForm name="startTime" />
+              </FormSection>
+              <FormSection title="End Time">
+                <TimeForm name="endTime" />
+              </FormSection>
+            </div>
+            <FormSection title="Description">
+              {getValues('description') && (
+                <DescriptionForm name="description" />
               )}
-            </div>
-          </div>
-          <div className="flex gap-6">
-            <div>
-              <Label>Start Time</Label>
-              <Controller
-                render={({ field }) => (
-                  <DateTimePickerDemo
-                    onChange={field.onChange}
-                    defaultValue={field.value}
-                  />
-                )}
-                name="startTime"
-                control={control}
-              />
-              {errors.startTime && (
-                <div className="flex items-center gap-1 text-xs text-red-500">
-                  <PiWarningBold />
-                  {errors.startTime.message as string}
-                </div>
-              )}
-            </div>
-            <div>
-              <Label>End Time</Label>
-              <Controller
-                render={({ field }) => (
-                  <DateTimePickerDemo
-                    onChange={field.onChange}
-                    defaultValue={field.value}
-                  />
-                )}
-                name="endTime"
-                control={control}
-              />
-              {errors.endTime && (
-                <div className="flex items-center gap-1 text-xs text-red-500">
-                  <PiWarningBold />
-                  {errors.endTime.message as string}
-                </div>
-              )}
-            </div>
-          </div>
-
-          <div className="flex flex-col gap-1">
-            <Label>Description</Label>
-            {getValues('description') && (
-              <Controller
-                render={({ field }) => (
-                  <TextEditor
-                    placeholder="Enter a description..."
-                    onChange={field.onChange}
-                    defaultValue={field.value}
-                  />
-                )}
-                name="description"
-                control={control}
-              />
-            )}
-            {errors.description && (
-              <div className="flex items-center gap-1 text-xs text-red-500">
-                <PiWarningBold />
-                required
-              </div>
-            )}
-          </div>
-
-          <div className="flex flex-col gap-1">
-            <div className="flex items-center gap-2">
-              <Label required={false}>Register Code</Label>
-              <Switch
-                onCheckedChange={() => {
-                  setShowInvitationCode(!showInvitationCode)
-                  setValue('invitationCode', null)
-                }}
-                checked={showInvitationCode}
-                className="data-[state=checked]:bg-black data-[state=unchecked]:bg-gray-300"
-              />
-            </div>
-            {showInvitationCode && (
-              <Input
-                id="invitationCode"
-                placeholder="Enter a Invitation Code"
+            </FormSection>
+            {getValues('invitationCode') && (
+              <SwitchField
+                name="invitationCode"
+                title="Invitation Code"
                 type="number"
-                className={cn(
-                  inputStyle,
-                  'w-[380px]',
-                  '[appearance:textfield] [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none'
-                )}
-                {...register('invitationCode')}
+                isInput={true}
+                placeholder="Enter a invitation code"
+                hasValue={showInvitationCode}
               />
             )}
-            {errors.invitationCode && (
-              <div className="flex items-center gap-1 text-xs text-red-500">
-                <PiWarningBold />
-                {errors.invitationCode.message as string}
-              </div>
-            )}
-          </div>
 
-          <div className="flex flex-col gap-1">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <Label>Contest Problem List</Label>
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <button>
-                      <MdHelpOutline className="text-gray-400 hover:text-gray-700" />
-                    </button>
-                  </PopoverTrigger>
-                  <PopoverContent
-                    side="top"
-                    className="mb-2 w-[680px] bg-black px-4 py-2 text-white"
-                  >
-                    <ul className="text-xs font-normal">
-                      <li>
-                        The problems in the contest problem list are initially
-                        set to &apos;not visible&apos; at the time of creating
-                        the contest
-                      </li>
-                      <li>
-                        They become visible according to the specified start
-                        time and remain inaccessible in the problem list
-                      </li>
-                      <li>
-                        throughout the duration of the contest. After the
-                        contest period ends, they become visible again in the
-                        problem list.
-                      </li>
-                    </ul>
-                  </PopoverContent>
-                </Popover>
+            <div className="flex flex-col gap-1">
+              <div className="flex items-center justify-between">
+                <ContestProblemListLabel />
+                <ImportProblemButton disabled={isLoading} isCreatePage={true} />
               </div>
-              <Button
-                type="button"
-                className="flex h-[36px] w-36 items-center gap-2 px-0"
-                disabled={isLoading}
-                onClick={() => {
-                  const formData = {
-                    title: getValues('title'),
-                    startTime: getValues('startTime'),
-                    endTime: getValues('endTime'),
-                    description: getValues('description'),
-                    invitationCode: getValues('invitationCode')
-                  }
-                  localStorage.setItem(
-                    'contestFormData',
-                    JSON.stringify(formData)
-                  )
-                  router.push('/admin/problem?import=true')
-                }}
-              >
-                <PlusCircleIcon className="h-4 w-4" />
-                <div className="mb-[2px] text-sm">Import Problem</div>
-              </Button>
+              <DataTableAdmin
+                // eslint-disable-next-line
+                columns={columns as any[]}
+                data={problems as ContestProblem[]}
+                enableDelete={true}
+                enableSearch={true}
+              />
             </div>
-            <DataTableAdmin
-              // eslint-disable-next-line
-              columns={columns as any}
-              data={problems}
-              enableDelete={true}
-              enableSearch={true}
-            />
-          </div>
-          <Button
-            type="submit"
-            className="flex h-[36px] w-[100px] items-center gap-2 px-0"
-            disabled={isLoading || isCreating}
-          >
-            <IoMdCheckmarkCircleOutline fontSize={20} />
-            <div className="mb-[2px] text-base">Create</div>
-          </Button>
+            <Button
+              type="submit"
+              className="flex h-[36px] w-[100px] items-center gap-2 px-0"
+              disabled={isLoading || isCreating}
+            >
+              <IoMdCheckmarkCircleOutline fontSize={20} />
+              <div className="mb-[2px] text-base">Create</div>
+            </Button>
+          </FormProvider>
         </form>
       </main>
     </ScrollArea>
