+++ apps/frontend/app/admin/utils.ts
@@ -1,0 +1,2 @@
+export const inputStyle =
+  'border-gray-200 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-gray-950'
