+++ apps/frontend/components/DataTableAdmin.tsx
@@ -166,7 +166,7 @@ export function DataTableAdmin<TData, TValue>({
         `importProblems-${contestId}`,
         JSON.stringify(problems)
       )
-      router.push(`/admin/contest/${contestId}`)
+      router.push(`/admin/contest/${contestId}/edit`)
     }
   }
